rewrites README and reads yaml of Proust's questionnaire
