+++ 2_adventures/001/sbartek/06/README.md
@@ -1,7 +1,6 @@
 # Questionnaire of Prost
 
-Here we start implement classes as described in 04.
+We continue with the project. We have create questionnaire from yaml
+file.
 
-I'm going to play with minitest, which I don't know.  I'll start with
-specs written in the same file where the classes are, and when the
-code ans specs grows I will refactor creating new files/directories.
+Next step will be answering the questionnaire.
