+++ 2_adventures/001/sbartek/06/proust.yaml
@@ -1,0 +1,81 @@
+- sentence: Your favorite virtue	
+  interpretation: The principal aspect of my personality	
+  answer: The need to be loved; more precisely, the need to be caressed and spoiled much more than the need to be admired
+- sentence: Your favorite qualities in a man.
+  interpretation: The quality that I desire in a man.	
+  answer: Manly virtues, and frankness in friendship.
+- sentence: Your favorite qualities in a woman.
+  interpretation: The quality that I desire in a woman.	
+  answer: Feminine charms
+- sentence: What you appreciate the most in your friends
+  interpretation: What I appreciate most about my friends.	
+  answer: To have tenderness for me, if their personage is exquisite enough to render quite high the price of their tenderness
+- sentence: Your main fault	
+  interpretation: My main fault	
+  answer: Not knowing, not being able to "want".
+- sentence: Your favourite occupation.	
+  interpretation: My favorite occupation.	
+  answer: Loving.
+- sentence: Your idea of happiness	
+  interpretation: My dream of happiness.	
+  answer: I am afraid it be not great enough, I dare not speak it, I am afraid of destroying it by speaking it.
+- sentence: Your idea of misery.
+  interpretation: What would be my greatest misfortune?	
+  answer: Not to have known my mother or my grandmother.
+- sentence: If not yourself, who would you be?	
+  interpretation: What I should like to be.	
+  answer: Myself, as the people whom I admire would like me to be.
+- sentence: Where would you like to live?
+  interpretation: The country where I should like to live. 
+  answer: A country where certain things that I should like would come true as though by magic, and where tenderness would always be reciprocated
+- sentence: Your favourite colour.	
+  interpretation: My favourite colour.	
+  answer: The beauty is not in the colours, but in their harmony.
+- sentence: Your favorite prose authors.	
+  interpretation: My favorite prose authors.	
+  answer: Currently, Anatole France and Pierre Loti.
+- sentence: Your favorite poets.	
+  interpretation: My favorite poets.
+  answer: Baudelaire and Alfred de Vigny.
+- sentence: Your favorite heroes in fiction.	
+  interpretation: My heroes in fiction.	
+  answer: Hamlet.
+- sentence: Your favorite heroines in fiction.	
+  interpretation: My favorite heroines in fiction.	
+  answer: Bérénice.
+- sentence: Your favorite painters and composers.	
+  interpretation: My favorite composers.	
+  answer: Beethoven, Wagner, Schumann.
+- sentence: Your heroes in real life.	
+  interpretation: My heroes in real life.	
+  answer: Mr. Darlu, Mr. Boutroux.
+- sentence: Your heroines in World history	
+  interpretation:  My heroines in history.	
+  answer: Cleopatra.
+- sentence: Your favorite names.	
+  interpretation:  My favorite names.
+  answer: I only have one at a time.
+- sentence: What I hate the most.	
+  interpretation: What I hate most of all.	
+  answer: What is bad about me.
+- sentence: World history characters I hate the most
+  interpretation: Historical figures that I despise the most.
+  answer: I am not educated enough.
+- sentence: The military event I admire the most	
+  interpretation: The military event that I admire most.	
+  answer: My military service!
+- sentence: "The natural talent I'd like to be gifted with"	  
+  interpretation: The gift of nature that I would like to have.	
+  answer: Will-power, and seductiveness.
+- sentence: How I wish to die	
+  interpretation: How I want to die.	
+  answer: Improved—and loved.
+- sentence: What is your present state of mind.	
+  interpretation: My present state of mind.	
+  answer: Boredom from having thought about myself to answer all these questions.
+- sentence: For what fault have you most toleration?	
+  interpretation: Faults for which I have the most indulgence.	
+  answer: Those that I understand.
+- sentence: Your favorite motto.	
+  interpretation: My motto.	
+  answer: I should be too afraid that it bring me misfortune.
